Moved tests to infrastructure package
